+++ OthelloSolverMultithreaded/OthelloSolverMultithreaded.h
@@ -34,4 +34,4 @@ extern COthelloSolverMultithreadedApp theApp;
 #define WM_CUSTOM_UPDATE_STATUS  (WM_USER + 1)
 #define WM_CUSTOM_SOLVER_DONE    (WM_USER + 2)
 
-#define APP_VERSION              "1.1.0"
+#define APP_VERSION              "1.2.0"